requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
scikit-learn==1.3.2
kneed==0.8.5
sqlalchemy==2.0.23
python-dotenv==1.0.0
pg8000==1.30.3 